refactor: post and get Inspectors are now a single endpoint
