+++ packages/backend/api/inspector/postProcessInspector.go
@@ -1,4 +1,4 @@
-package postInspector
+package inspector
 
 import (
 	"strings"
