+++ packages/backend/api/inspector/test_inspector/GetBaseThreshold_test.go
@@ -1,11 +1,11 @@
-package getRecentTicketInspectorInfo_test
+package test_inspector
 
 import (
 	"math"
 	"testing"
 	"time"
 
-	"github.com/FreiFahren/backend/api/getRecentTicketInspectorInfo"
+	"github.com/FreiFahren/backend/api/inspector"
 )
 
 func TestBaseThreshold(t *testing.T) {
@@ -29,7 +29,7 @@ func TestBaseThreshold(t *testing.T) {
 	for _, tc := range testCases {
 		t.Run(tc.dateTime, func(t *testing.T) {
 			currentTime, _ := time.Parse(layout, tc.dateTime)
-			result := getRecentTicketInspectorInfo.GetBaseThreshold(currentTime.Hour(), currentTime.Minute(), currentTime)
+			result := inspector.GetBaseThreshold(currentTime.Hour(), currentTime.Minute(), currentTime)
 			roundedResult := math.Round(result*100) / 100
 			if roundedResult != tc.expected {
 				t.Errorf("BaseThreshold(%s) = %v; want %v", tc.dateTime, roundedResult, tc.expected)
