+++ packages/backend/docs/swagger.yaml
@@ -116,38 +116,7 @@ info:
   title: FreiFahren API Documentation
   version: "1.0"
 paths:
-  /basics/newInspector:
-    post:
-      consumes:
-      - application/json
-      description: |-
-        Accepts a JSON payload with details about a ticket inspector's current location.
-        This endpoint validates the provided data, processes necessary computations for linking stations and lines,
-        inserts the data into the database, and triggers an update to the risk model used in operational analysis.
-      parameters:
-      - description: Data about the inspector's location and activity
-        in: body
-        name: inspectorData
-        required: true
-        schema:
-          $ref: '#/definitions/utils.InspectorRequest'
-      produces:
-      - application/json
-      responses:
-        "200":
-          description: Successfully processed and inserted the inspector data with
-            computed linkages and risk model updates.
-          schema:
-            $ref: '#/definitions/utils.ResponseData'
-        "400":
-          description: 'Bad Request: Missing or incorrect parameters provided.'
-        "500":
-          description: 'Internal Server Error: Error during data processing or database
-            insertion.'
-      summary: Submit ticket inspector data
-      tags:
-      - basics
-  /basics/recent:
+  /basics/inspectors:
     get:
       consumes:
       - application/json
@@ -181,6 +150,36 @@ paths:
           schema:
             type: string
       summary: Retrieve information about recent ticket inspector reports
+      tags:
+      - basics
+    post:
+      consumes:
+      - application/json
+      description: |-
+        Accepts a JSON payload with details about a ticket inspector's current location.
+        This endpoint validates the provided data, processes necessary computations for linking stations and lines,
+        inserts the data into the database, and triggers an update to the risk model used in operational analysis.
+      parameters:
+      - description: Data about the inspector's location and activity
+        in: body
+        name: inspectorData
+        required: true
+        schema:
+          $ref: '#/definitions/utils.InspectorRequest'
+      produces:
+      - application/json
+      responses:
+        "200":
+          description: Successfully processed and inserted the inspector data with
+            computed linkages and risk model updates.
+          schema:
+            $ref: '#/definitions/utils.ResponseData'
+        "400":
+          description: 'Bad Request: Missing or incorrect parameters provided.'
+        "500":
+          description: 'Internal Server Error: Error during data processing or database
+            insertion.'
+      summary: Submit ticket inspector data
       tags:
       - basics
   /data/id:
