+++ packages/backend/main.go
@@ -6,12 +6,11 @@ import (
 	"github.com/FreiFahren/backend/Rstats"
 	"github.com/FreiFahren/backend/api/getAllStationsAndLines"
 	"github.com/FreiFahren/backend/api/getId"
-	"github.com/FreiFahren/backend/api/getRecentTicketInspectorInfo"
 	"github.com/FreiFahren/backend/api/getSegmentColors"
 	"github.com/FreiFahren/backend/api/getStationDistance"
 	"github.com/FreiFahren/backend/api/getStationName"
 	"github.com/FreiFahren/backend/api/getStats"
-	"github.com/FreiFahren/backend/api/postInspector"
+	"github.com/FreiFahren/backend/api/inspector"
 	"github.com/FreiFahren/backend/data"
 	"github.com/FreiFahren/backend/database"
 	_ "github.com/FreiFahren/backend/docs"
@@ -112,28 +111,19 @@ func main() {
 	// Ensure the required table exists
 	database.CreateTicketInfoTable()
 
-	// Post a new ticket inspector
-	apiHOST.POST("/basics/newInspector", postInspector.PostInspector)
+	apiHOST.POST("/basics/inspectors", inspector.PostInspector)
+	apiHOST.GET("/basics/inspectors", inspector.GetTicketInspectorsInfo)
 
-	// Return the recent ticket inspector info
-	apiHOST.GET("/basics/recent", getRecentTicketInspectorInfo.GetRecentTicketInspectorInfo)
-
-	// Return the name for given id
 	apiHOST.GET("/data/station", getStationName.GetStationName)
 
-	// Return all stations with their id
 	apiHOST.GET("/data/list", getAllStationsAndLines.GetAllStationsAndLines)
 
-	// Return the id for given name
 	apiHOST.GET("/data/id", getId.GetStationId)
 
-	// Return the distance between two stations
 	apiHOST.GET("/transit/distance", getStationDistance.GetStationDistance)
 
-	// Get usage statistics
 	apiHOST.GET("/statistics/stats", getStats.GetStats)
 
-	// Get the list of highlighted segments and their colors
 	apiHOST.GET("/risk-prediction/getSegmentColors", getSegmentColors.GetSegmentColors)
 
 	apiHOST.Start(":8080")
